+++ httpsig/utils.py
@@ -4,6 +4,47 @@ import hashlib
 import base64
 
 from Crypto.PublicKey import RSA
+from Crypto.Hash import SHA, SHA256, SHA512
+
+ALGORITHMS = frozenset(['rsa-sha1', 'rsa-sha256', 'rsa-sha512', 'hmac-sha1', 'hmac-sha256', 'hmac-sha512'])
+HASHES = {'sha1':   SHA,
+          'sha256': SHA256,
+          'sha512': SHA512}
+
+class HttpSigException(Exception):
+    pass
+
+def generate_message(required_headers, headers, host=None, method=None, path=None):
+    headers = CaseInsensitiveDict(headers)
+    
+    if not required_headers:
+        required_headers = ['date']
+    
+    signable_list = []
+    for h in required_headers:
+        if h == '(request-line)':
+            if not method or not path:
+                raise Exception('method and path arguments required when using "(request-line)"')
+            signable_list.append('%s: %s %s' % (h, method.lower(), path))
+
+        elif h == 'host':
+            # 'host' special case due to requests lib restrictions
+            # 'host' is not available when adding auth so must use a param
+            # if no param used, defaults back to the 'host' header
+            if not host:
+                if 'host' in headers:
+                    host = headers[h]
+                else:
+                    raise Exception('missing required header "%s"' % (h))
+            signable_list.append('%s: %s' % (h.lower(), host))
+        else:
+            if h not in headers:
+                raise Exception('missing required header "%s"' % (h))
+
+            signable_list.append('%s: %s' % (h.lower(), headers[h]))
+
+    signable = '\n'.join(signable_list)
+    return signable
 
 def lkv(d):
     parts = []